backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
passlib[bcrypt]
python-jose
pydantic
httpx
python-multipart
python-dotenv
